requirements.txt
# Core Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# HTTP Requests and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI and Language Models - Multi-Provider Support
openai>=1.30.0
anthropic>=0.30.0
google-generativeai>=0.5.0

# OCR and Image Processing
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.24.0

# Optional OCR (fallback)
# pytesseract>=0.3.10
# cairosvg>=2.7.0

# Web Browser Automation (Playwright replaces Selenium)
playwright>=1.40.0

# PDF Processing
PyPDF2>=3.0.0

# Sigma Rule Processing
pysigma>=0.10.0
pyyaml>=6.0.0

# Terminal UI and Display
rich>=13.0.0

# Natural Language Processing
nltk>=3.8.0

# Data Visualization (for PDF reports)
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# PDF Report Generation
reportlab>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Data Types and Utilities
typing-extensions>=4.8.0

# Encryption (session management)
cryptography>=41.0.0
